fix generation test
