+++ src/blop/tests/ax/test_generation.py
@@ -4,43 +4,17 @@ Tests the public API of LatentKernel - a Matérn kernel with learned affine
 transformation using SO(N) parameterization for orthogonal rotations.
 """
 
-import numpy as np
 import pytest
 from bluesky.run_engine import RunEngine
 
 from blop.ax.agent import Agent
-from blop.ax.dof import DOFConstraint, RangeDOF
+from blop.ax.dof import RangeDOF
+from blop.ax.generation import get_generation_strategy
 from blop.ax.objective import Objective
+from blop.evaluation.test_functions import TestFunctionEvaluation
 
-from ..conftest import MovableSignal, ReadableSignal
+from ..conftest import MovableSignal
 from ..test_plans import _collect_optimize_events
-
-
-class TestFunctionEvaluation:
-    def __init__(self, tiled_client):
-        self.tiled_client = tiled_client
-
-    def __call__(self, uid: str, suggestions: list[dict]) -> list[dict]:
-
-        outcomes = []
-
-        for suggestion in suggestions:
-            # Special key to identify a suggestion
-            suggestion_id = suggestion["_id"]
-            x1 = suggestion["x1"]
-            x2 = suggestion["x2"]
-
-            outcomes.append(
-                {
-                    "test_function_1": 1 - np.exp(-((x1 - 2 * x2 - 1) ** 2) - 1e-3 * (2 * x1 + x2 - 0.5) ** 2),
-                    "test_function_2": 1 - np.exp(-1e-3 * (x1 - 2 * x2 - 1) ** 2 - (2 * x1 + x2 - 0.5) ** 2),
-                    "_id": suggestion_id,
-                }
-            )
-
-            # outcomes.append({"test_function": (x1 ** 2 + x2 - 11) ** 2 + (x1 + x2 ** 2 - 7) ** 2, "_id": suggestion_id})
-
-        return outcomes
 
 
 @pytest.fixture(scope="function")
@@ -48,22 +22,26 @@ def RE():
     return RunEngine({})
 
 
-def test_optimize(RE, mock_acquisition_plan, mock_evaluation_function):
-    movable1 = MovableSignal(name="test_movable1")
-    movable2 = MovableSignal(name="test_movable2")
-    readable = ReadableSignal(name="test_readable")
-    dof1 = RangeDOF(actuator=movable1, bounds=(0, 10), parameter_type="float")
-    dof2 = RangeDOF(actuator=movable2, bounds=(0, 10), parameter_type="float")
-    constraint = DOFConstraint(constraint="x1 + x2 <= 10", x1=dof1, x2=dof2)
-    objective = Objective(name="test_objective", minimize=False)
+def test_optimize(RE):
+
+    dofs = [
+        RangeDOF(actuator=MovableSignal("x1"), bounds=(-2.0, 2.0), parameter_type="float"),
+        RangeDOF(actuator=MovableSignal("x2"), bounds=(-2.0, 2.0), parameter_type="float"),
+    ]
+
+    objectives = [
+        Objective(name="fitness", minimize=False),
+    ]
+
     agent = Agent(
-        sensors=[readable],
-        dofs=[dof1, dof2],
-        objectives=[objective],
-        evaluation_function=mock_evaluation_function,
-        dof_constraints=[constraint],
-        acquisition_plan=mock_acquisition_plan,
-        name="test_experiment",
+        sensors=[],
+        dofs=dofs,
+        objectives=objectives,
+        evaluation_function=TestFunctionEvaluation(),
+        name="test",
+        description="Optimization on a test function",
+        experiment_type="demo",
+        generation_strategy=get_generation_strategy("beamline"),
     )
 
     callback, events = _collect_optimize_events()
